⚡️ performance: 아코디언 아이템 클릭시 전체 리렌더링 개선
