+++ src/components/Accordion/accordion.scss
@@ -17,15 +17,12 @@
       aspect-ratio: 1/1;
       transition: all var(--transition-duration);
     }
-    &.act::after {
-      transform: rotate(180deg);
-    }
-
     .badge {
       vertical-align: 0.1em;
       margin-left: 0.6rem;
     }
   }
+
   .accordion-body {
     display: none;
     padding: var(--title-extra) 0;
@@ -33,12 +30,18 @@
     color: var(--color-gray-100);
     font-size: var(--title-medium);
     word-break: keep-all;
-
-    &.act {
-      display: block;
-    }
     pre {
       white-space: pre-line;
+    }
+  }
+
+  // 활성화
+  .accordion-item.active {
+    .accordion-header::after {
+      transform: rotate(180deg);
+    }
+    .accordion-body {
+      display: block;
     }
   }
